+++ src/Module/Application/Stream/RandomAction.php
@@ -1,8 +1,9 @@
 <?php
+
 /*
  * vim:set softtabstop=4 shiftwidth=4 expandtab:
  *
- *  LICENSE: GNU Affero General Public License, version 3 (AGPL-3.0-or-later)
+ * LICENSE: GNU Affero General Public License, version 3 (AGPL-3.0-or-later)
  * Copyright 2001 - 2022 Ampache.org
  *
  * This program is free software: you can redistribute it and/or modify
@@ -22,58 +23,47 @@
 
 declare(strict_types=0);
 
-namespace Ampache\Module\Application\DemocraticPlayback;
+namespace Ampache\Module\Application\Stream;
 
 use Ampache\Config\ConfigContainerInterface;
 use Ampache\Config\ConfigurationKeyEnum;
-use Ampache\Repository\Model\Democratic;
-use Ampache\Module\Application\ApplicationActionInterface;
-use Ampache\Module\Application\Exception\AccessDeniedException;
-use Ampache\Module\Authorization\AccessLevelEnum;
 use Ampache\Module\Authorization\GuiGatekeeperInterface;
-use Ampache\Module\Util\Ui;
-use Ampache\Module\Util\UiInterface;
+use Ampache\Repository\Model\Random;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Log\LoggerInterface;
 
-final class ManageAction implements ApplicationActionInterface
+final class RandomAction extends AbstractStreamAction
 {
-    public const REQUEST_KEY = 'manage';
+    public const REQUEST_KEY = 'random';
 
-    private UiInterface $ui;
+    private LoggerInterface $logger;
 
     private ConfigContainerInterface $configContainer;
 
     public function __construct(
-        UiInterface $ui,
+        LoggerInterface $logger,
         ConfigContainerInterface $configContainer
     ) {
-        $this->ui              = $ui;
+        $this->logger          = $logger;
         $this->configContainer = $configContainer;
+
+        parent::__construct($logger, $configContainer);
     }
 
     public function run(ServerRequestInterface $request, GuiGatekeeperInterface $gatekeeper): ?ResponseInterface
     {
-        /* Make sure they have access to this */
-        if (
-            $this->configContainer->isFeatureEnabled(ConfigurationKeyEnum::ALLOW_DEMOCRATIC_PLAYBACK) === false ||
-            $gatekeeper->mayAccess(AccessLevelEnum::TYPE_INTERFACE, AccessLevelEnum::LEVEL_MANAGER) === false
-        ) {
-            throw new AccessDeniedException();
+        if ($this->preCheck($gatekeeper) === false) {
+            return null;
         }
+        $randomId   = (int) $request->getQueryParams()['random_id'];
+        $randomType = $request->getQueryParams()['random_type'];
+        $urls       = [Random::get_play_url($randomType, $randomId)];
 
-        $this->ui->showHeader();
-
-        $democratic = Democratic::get_current_playlist();
-        $democratic->set_parent();
-        $democratic->format();
-
-        // Show the create page
-        require_once Ui::find_template('show_edit_democratic.inc.php');
-
-        $this->ui->showQueryStats();
-        $this->ui->showFooter();
-
-        return null;
+        return $this->stream(
+            [],
+            $urls,
+            $this->configContainer->get(ConfigurationKeyEnum::PLAY_TYPE)
+        );
     }
 }