+++ src/Module/Playback/Stream_Playlist.php
@@ -316,9 +316,6 @@ class Stream_Playlist
                     $url['image_url'] = Art::url($object->podcast, 'podcast', $api_session, (AmpConfig::get('ajax_load') ? 3 : 4));
                     $url['codec']     = $object->type;
                     break;
-                case 'random':
-                    $url['title'] = 'Random URL';
-                    break;
                 default:
                     $url['title'] = Stream_Url::get_title($url['url']);
                     $url['time']  = -1;