+++ src/Module/System/Core.php
@@ -311,6 +311,9 @@ class Core
      */
     public static function is_readable($path)
     {
+        if (!$path) {
+            return false;
+        }
         if (file_exists($path)) {
             if (is_dir($path)) {
                 $handle = opendir($path);
@@ -342,7 +345,7 @@ class Core
      */
     public static function get_filesize($filename)
     {
-        if (!file_exists($filename)) {
+        if (!$filename || !file_exists($filename)) {
             return 0;
         }
         $size = filesize($filename);